BAM-1381 use wildcard descendant locator for test quickfix
These tests actually weren't broken by an update of our online content, but
by an update of a dependency, which slightly changed the behavior of the
locator to consider the text of the selected element itself, not including
that of its children.

We simply change the locator to also match all descendants.